Create lifeDomain
This was the main prep page. This domain only uses cleric spells for its unique domain spells so the other domain windows were a copy of this one with the unique spells being the only change between each domain window. This is also the main page with comments on the different classes and certain actions of buttons.